I would like to add a new section to the Entity page that displays the e
